feat: refactor spd matching logic for poet + custom benchmarks
